VcfBreakMulti implementation finished. Moved VcfCheck functionality into VcfFile class to make it usable in class/tool tests as well.
