+++ src/VcfAnnotateFromBed/main.cpp
@@ -3,6 +3,7 @@
 #include "Helper.h"
 #include "BedFile.h"
 #include "ChromosomalIndex.h"
+#include "VcfFile.h"
 #include <QFile>
 #include <QSharedPointer>
 
@@ -81,7 +82,7 @@ public:
 
 			//split line and extract variant infos
 			QList<QByteArray> parts = line.split('\t');
-			if (parts.count()<8) THROW(FileParseException, "VCF with too few columns: " + line);
+			if (parts.count()<VcfFile::MIN_COLS) THROW(FileParseException, "VCF line with too few columns: " + line);
 			Chromosome chr = parts[0];
 			bool ok = false;
 			int start = parts[1].toInt(&ok);
