+++ src/tools-TEST/VcfAnnotateFromBed_Test.h
@@ -1,4 +1,5 @@
 #include "TestFramework.h"
+#include "TestFrameworkNGS.h"
 #include "Settings.h"
 
 TEST_CLASS(VcfAnnotateFromBed_Test)
@@ -11,6 +12,7 @@ private slots:
 		EXECUTE("VcfAnnotateFromBed", "-in " + TESTDATA("data_in/VcfAnnotateFromBed_in1.vcf") + " -out out/VcfAnnotateFromBed_out1.vcf -name OMIM -bed " + TESTDATA("data_in/VcfAnnotateFromBed_in1.bed") );
 		REMOVE_LINES("out/VcfAnnotateFromBed_out1.vcf", QRegExp("##INFO=<ID=OMIM"))
 		COMPARE_FILES("out/VcfAnnotateFromBed_out1.vcf", TESTDATA("data_out/VcfAnnotateFromBed_out1.vcf"));
+		VCF_IS_VALID("out/VcfAnnotateFromBed_out1.vcf")
 	}
 
 };
