+++ src/tools-TEST/VcfLeftNormalize_Test.h
@@ -1,5 +1,6 @@
 #include "TestFramework.h"
 #include "Settings.h"
+#include "TestFrameworkNGS.h"
 
 TEST_CLASS(VcfLeftNormalize_Test)
 {
@@ -13,6 +14,7 @@ private slots:
 
 		EXECUTE("VcfLeftNormalize", "-in " + TESTDATA("data_in/VcfLeftNormalize_in1.vcf") + " -out out/VcfLeftNormalize_out1.vcf -ref " + ref_file);
 		COMPARE_FILES("out/VcfLeftNormalize_out1.vcf", TESTDATA("data_out/VcfLeftNormalize_out1.vcf"));
+		VCF_IS_VALID("out/VcfLeftNormalize_out1.vcf")
 	}
 
 };
