+++ src/tools-TEST/VcfStreamSort_Test.h
@@ -1,4 +1,5 @@
 #include "TestFramework.h"
+#include "TestFrameworkNGS.h"
 #include "Settings.h"
 
 TEST_CLASS(VcfStreamSort_Test)
@@ -10,6 +11,8 @@ private slots:
 	{
 		EXECUTE("VcfStreamSort", "-n 4 -in " + TESTDATA("data_in/VcfStreamSort_in1.vcf") + " -out out/VcfStreamSort_out1.vcf");
 		COMPARE_FILES("out/VcfStreamSort_out1.vcf", TESTDATA("data_out/VcfStreamSort_out1.vcf"));
+
+		VCF_IS_VALID("out/VcfStreamSort_out1.vcf")
 	}
 
 };
