+++ src/tools-TEST/data_in/VcfSort_in3.vcf
@@ -1,8 +1,8 @@
+##fileformat=VCFv4.1
+##fileDate=20160408
 ##cmdline=/mnt/SRV017/share/opt/strelka_workflow-1.0.14/libexec/consolidateResults.pl --config=/tmp/vc_strelka_341807/strelkaAnalysis/config/run.config.ini
 ##content=strelka somatic indel calls
 ##content=strelka somatic snv calls
-##fileDate=20160408
-##fileformat=VCFv4.1
 ##filter="Repeat & iHpol & BCNoise & QSI_ref & SpanDel & QSS_ref"
 ##germlineIndelTheta=0.0001
 ##germlineSnvTheta=0.001
