+++ src/tools-TEST/tools-TEST.pro
@@ -96,7 +96,7 @@ HEADERS += \
     FastqAddBarcode_Test.h \
     BamFilter_Test.h \
     VcfCheck_Test.h \
-    VcfMultiBreak_Test.h
+    VcfBreakMulti_Test.h
 
 SOURCES += \
     main.cpp \
